# projects/backend/app/schemas/token.py
from pydantic import UUID4, BaseModel


class Token(BaseModel):
    access_token: str
    token_type: str


class TokenData(BaseModel):
    id: UUID4
    role: str = None
    account_id: UUID4 = None